+++ schemars/src/json_schema_impls/mod.rs
@@ -61,10 +61,7 @@ mod ffi;
 mod arrayvec07;
 
 #[cfg(feature = "bytes1")]
-mod bytes1 {
-    forward_impl!(bytes1::Bytes => alloc::vec::Vec<u8>);
-    forward_impl!(bytes1::BytesMut => alloc::vec::Vec<u8>);
-}
+mod bytes1;
 
 #[cfg(feature = "chrono04")]
 mod chrono04;