+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - ⚠️ MSRV is now 1.70 ⚠️
 
+### Fixed
+
+- The "deserialize" schema for `bytes::Bytes`/`BytesMut` now allows strings, matching the actual deserialize behaviour of the types.
+
 ## [1.0.0-alpha.15] - 2024-09-05
 
 ### Added